Big update + refactoring start
Scrapping event system to allow for better ecs intergration, key input will be handled by the application storing an array of keys from the current frame and previous frame. Each stored key has 3 values bits 0 -> 2, pressed, released and repeat which are all determined by the system. Linux threads has new functionality for termination and failing to create a thread will allow the program to carry on but execute the command on the main thread.