Mage overhaul
Lots to cover, the old api was a mess, half finished systems and bad practices whilst small would have been a real pain to fix. Using a single header was a bad idea due to bad compile times on slower systems when developing. The new project aims to be dependency free without 3rd party libraries to allow for the GNU GPL3 license without any issues. Some code will be recycled including some of the helper scripts and vulkan code but alot is being built from the ground up with usablility on the forefront of design princibles. Unit tests will be a major focus which mage had lacked previously.